Update and rename msgbox_v1.0.py to msgbox_v1.2.py
Prepended private functions with an underscore '_' for clarity. 
Added improved handling of user errors. If user indicates buttons that are not allowed, it will default to 'OK' instead of failing. 
Added '_ezmsg' feature allowing lazy usage in which all args are placed in the first string argument, separated by commas.
Changed testing to use the "if __name__ == '__main__' so that they run when the file is run directly, but not when imported as a module.